tg-bot/requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
psutil==5.9.6
pydantic
supabase==2.9.1
requests==2.31.0